// apps/web/src/api/generated/schemas/responseEnvelopeListSaleRead.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseEnvelopeListSaleReadData } from './responseEnvelopeListSaleReadData';
import type { ResponseEnvelopeListSaleReadMessage } from './responseEnvelopeListSaleReadMessage';
import type { ResponseEnvelopeListSaleReadError } from './responseEnvelopeListSaleReadError';
import type { ResponseEnvelopeListSaleReadMeta } from './responseEnvelopeListSaleReadMeta';
import type { ResponseEnvelopeListSaleReadRequestId } from './responseEnvelopeListSaleReadRequestId';

export interface ResponseEnvelopeListSaleRead {
  success?: boolean;
  data?: ResponseEnvelopeListSaleReadData;
  message?: ResponseEnvelopeListSaleReadMessage;
  error?: ResponseEnvelopeListSaleReadError;
  meta?: ResponseEnvelopeListSaleReadMeta;
  requestId?: ResponseEnvelopeListSaleReadRequestId;
  timestamp?: string;
}